Pagination for query from form in header, separate handlers for pagination
