+++ src/index.js
@@ -7,5 +7,7 @@ import './js/index.js';
 import getMoviesByQuery from './js/requestsToServer/getMoviesByQuery';
 import getTrendingMovies from "./js/requestsToServer/getTrendingMovies"
 import getMovieById from './js/requestsToServer/getMovieById';
+import debouncedInputQuery from './js/homePage';
 
 getTrendingMovies();
+// getMoviesByQuery("Ten");
