+++ src/js/refsPaginat.js
@@ -3,6 +3,7 @@ const refsPaginate = {
   pagesList: document.querySelector(".js-pagination-pages"),
   btnPrev: document.querySelector(".js-pagination__btn--prev"),
   btnNext: document.querySelector(".js-pagination__btn--next"),
+  searchQuery: document.querySelector(".search-input"),
 }
 
 export default refsPaginate;
