+++ src/js/requestsToServer/getTrendingMovies.js
@@ -1,25 +1,16 @@
 import refsPaginate from '../refsPaginat';
-import handlePagination from '../handlePagination';
+import handlePaginationTrand from '../paginate/handlePaginationTrand';
 import fetchQueryService from '../services/fetchQueryService';
 import renderTemplateListMovies from '../renderTemplateListMovies';
 
 const getTrendingMovies = () => {
+  fetchQueryService.resetPage();
   fetchQueryService.fetchMoviesTrand()
     .then(response => {
-      handlePagination(response)
+      handlePaginationTrand(response)
     })
     .catch(error => console.log(error.message));
 
 }
-
-// const generateButtons = (countFilms) => {
-//   if (countFilms < 1) {
-//     return;
-//   }
-//   if (countFilms < 5 && countFilms > 1) {
-//
-//   }
-// }
-
 
 export default getTrendingMovies;
